requirements.txt
pyTelegramBotAPI==4.27.0
supabase==2.17.0
python-dotenv==1.0.1
httpx>=0.24,<1.0
